+++ router/delivery/http/router_handler.go
@@ -135,7 +135,7 @@ func (a *RouterHandler) GetOptimalQuote(c echo.Context) (err error) {
 		scalingFactor = a.getSpotPriceScalingFactor(tokenIn.Denom, tokenOutDenom)
 	}
 
-	_, _, err = quote.PrepareResult(ctx, scalingFactor)
+	_, _, err = quote.PrepareResult(ctx, scalingFactor, a.logger)
 	if err != nil {
 		return err
 	}
@@ -224,7 +224,7 @@ func (a *RouterHandler) GetDirectCustomQuote(c echo.Context) (err error) {
 		scalingFactor = a.getSpotPriceScalingFactor(tokenIn.Denom, tokenOutDenom[len(tokenOutDenom)-1])
 	}
 
-	_, _, err = quote.PrepareResult(ctx, scalingFactor)
+	_, _, err = quote.PrepareResult(ctx, scalingFactor, a.logger)
 	if err != nil {
 		return c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
 	}